+++ CHANGELOG.md
@@ -37,6 +37,7 @@ Ref: https://keepachangelog.com/en/1.0.0/
 
 ## v0.5.0
 
+* [#85](https://github.com/osmosis-labs/sqs/pull/85) Human readable denom params in router
 * [#84](https://github.com/osmosis-labs/sqs/pull/84) enable swagger
 * [#83](https://github.com/osmosis-labs/sqs/pull/83) /tokens/metadata/:denom endpoint